[KFI] Move files to right directory.
